pyproject.toml
[tool.poetry]
name = "digitalgaz"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.32.0"
fastapi = "^0.115.2"
python-dotenv = "^1.0.1"
aiofiles = "^24.1.0"
python-multipart = "^0.0.12"
ultralytics = "^8.3.14"
aiogram = "^3.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
web = "digitalgaz.main:web_start"
bot = "digitalgaz.main:bot_start"

